pyproject.toml
[tool.poetry]
name = "app"
version = "1.0.0"
description = ""
authors = ["Evgeny Bagirov <e.bagirovv@gmail.com>"]
readme = "README.md"

[tool.poetry.plugins.dotenv]
ignore = "false"
location = ".env.dev"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^2.1.3"
pandas = "^2.2.3"
nltk = "^3.9.1"
transformers = "^4.47.0"
scikit-learn = "^1.5.2"
torch = "^2.5.1"
torchmetrics = "^1.6.0"
datasets = "^3.1.0"
comet-ml = "^3.47.4"
python-dotenv = "^1.0.1"
fastapi = "^0.115.6"
torchvision = "^0.20.1"
uvicorn = "^0.32.1"
python-multipart = "^0.0.19"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
